Added BL and DL folders for Interface project, and renamed classes.
